refactor: action names
