+++ apps/web/src/actions/analytics/dashboard/DashboardAnalytics.ts
@@ -2,7 +2,7 @@
 
 import { GetDashboardCategories } from "@/actions/analytics/dashboard/GetDashboardCategories";
 import { GetDashboardStats } from "@/actions/analytics/dashboard/GetDashboardStats";
-import { GetLastTransactions } from "@/actions/analytics/getLastTransactions";
+import { GetLastTransactions } from "@/actions/analytics/dashboard/GetLastTransactions";
 
 export const DashboardAnalytics = async () => {
   const stats = await GetDashboardStats();
